feat(campus): tabbed right panel — Inspect / Tools / Overview
The right dock is now one tabbed panel instead of a single scrolling
Overview view:

- Inspect  — the selected element's actions (inline editing lands
             next commit); selecting an element auto-switches here
- Tools    — content-placement tools; Place POI as a tool tile
- Overview — campus stats grid + accessibility coverage

Replaces views/overview.tsx with views/panel.tsx (campusPanelView,
dock id "panel"). Built on the WorkflowTabBar DS primitive.
