+++ apps/campus/lib/workbench/views/index.ts
@@ -1,5 +1,5 @@
 export { mapView } from "./map";
-export { overviewView } from "./overview";
+export { campusPanelView } from "./panel";
 export { tableView } from "./table";
 export { hierarchyView } from "./hierarchy";
 export { aiPanelView } from "./ai-panel";
